Fixed failing spec and converted types up front.
I included two changes in this commit. First spec doesn't like treating something as a function and a vector (or a map). I split out the definitions to let me have both a function and a collection. Maybe there is a better way to do this.

Next, I decided to convert the seven-segments to ints earlier in the process. This means I'm working with smaller more readable data.